adding application cache
seems to work when viewing locally, i.e. on closing and re-opening the window the only request that hits the server is for the cache manifest file.

TODO: form the cache manifest file programmatically. right now every time we add a file we need to be careful to add it to the cache, which kinda sucks.