Draw topnav even when the user navigates to #css or #css/.
When the user loads dochub.io, the location is updated to #css/. If the
user then refreshed, previously there was no route to #css/ so the
topnav was never rendered and the homepage never appeared to finish
loading. With this change, the #css/:query route is more forgiving so it
will handle #css/ and #css.