Make the tocresult links have hrefs, and handler ignores click with modifiers.
By setting href, the search result links are now keyboard-accessible.
Also, the click handler now ignores right-click and Ctrl-click
(Cmd-click) so the user can open the link in a new tab.

In order to calculate the fragment, languageview must pass languageName
to searchresults, which passes it to mozdevcssprop. Similarly,
pagescrapedlanguageview sends languageName to pagesearchresults, which
sends it to mozdevcssprop to render..