Condense search window when there are fewer results
Currently the search window in condensed view (narrower than 700px or so) takes
up the entire height of the page, making it very difficult to scroll past as
scrollbars are nested inside of each other. By changing the height to
maxHeight, the search window will only take up the full page if there is more
than a full page of search results. As the user filters the search options, the
search window will shrink as well, displaying the results "above the fold".

Screenshot: https://skitch.com/kburke/en32p/dochub-instant-documentation-search